Upgrade bitcoin and miniscript
Upgrade to `miniscript` v0.10.0 and `bitcoin` v0.30.0

API Changes:

- `Network` is now `non_exhaustive` so we implement `TryFrom` instead of
`From`.
- `get_client`: Now takes an `HWIChain` instead of a generic, this is
less general, and less usable. Done because `From<Network>` can no
longer be implemented on `HWIChain` because of the newly added
`non_exhaustive` on `Network`.
- `Address` now has a tag and can only be deserialized unchecked. We
just call `assume_checked` in the unit tests.

Is the error handling for new `TryFrom` acceptable to the project?