Remove base64 dependency
We currently use `base64` to deserialize a vector. At the same time we
use `base64` indirectly from `rust-bitcoin` when deserializing a PSBT.

If we remove the dependency on `base64` and use the re-export from
bitcoin it saves having to keep the dependency and transient dependency
in sync so we don't get multiple versions of the same crate in the
dependency tree.