Rework the crate error type
Currently we are converting error types to strings, this makes the
library worse when used with standard error handling crates by
downstream users. Instead we can put each error type inside the
respective error variant and implement `std::error::Error` so that
`source` can be called by error handling libraries for better error
output.

Please note this removes `PartialEq` and `Eq` from the error type which
makes it harder to use. This is an ecosystem-wide problem caused by the
`io::Error`, there is currently (to the best of my knowledge) no good
solution other than removing the derives.

This is an API breaking change.