Use bitcoin::Network
The `HWIChain` type is only required so we can get around the orphan
rule and implement `IntoPy`. We can use a wrapper type (wrapping
`bitcoin::Network`) to achieve this and the use plain old `Network`
everywhere else. This simplifies the code and makes it more clear to
the casual reader who is familiar already with `rust-bitcoin`.