Bump MSRV to Rust 1.48.0
Recently the `rust-bitcoin` org bumped their MSRV to Rust 1.48, we can
do the same here.

`serde` v1.0.157 uses syn 2.0 which is edition 2021 so cannot be built
with 1.48 - we need to pin `serde` to v1.0.156

We can remove all the other pinning, features, and dependencies that
were required to get 1.41.1 to build.